feat: Add comprehensive location management system with GPS and search
- Implement LocationService with GPS → User Selected → Default (Bengaluru) fallback
- Add geocoding service using Open-Meteo and BigDataCloud APIs for city/pincode search
- Create location picker UI components for Android (BottomSheet) and iOS (Sheet)
- Add current location display in app header with tap-to-select functionality
- Implement proper offline/no internet state handling with specific error screens
- Add loading states for location detection and initial data fetch
- Update WeatherViewModel to handle location events (SelectLocation, RequestCurrentLocation, SearchLocations)
- Add comprehensive logging for location detection debugging
- Create cross-platform LocationManager interface with Android/iOS implementations
- Add location permissions handling and fallback mechanisms
- Update dependency injection with new location services
- Fix iOS build issues with inline location components

+++ androidApp/src/main/kotlin/com/weather/android/ui/screen/WeatherDetailScreen.kt
@@ -381,7 +381,7 @@ private fun getDetailCards(weather: Weather): List<DetailCardData> {
             icon = "💧",
             title = "Humidity",
             value = "${weather.humidity}%",
-            subtitle = getHumidityDescription(weather.humidity.toInt())
+            subtitle = getHumidityDescription(weather.humidity)
         ),
         DetailCardData(
             icon = WeatherUtils.getWeatherEmoji(weather.condition),
